backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==23.2.0
Flask-Session==0.5.0
python-dotenv==1.0.0
authlib==1.2.1
gunicorn==21.2.0 